+++ examples/resources/meraki_organization_adaptive_policy_groups/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_organization_adaptive_policy_groups.example "<organization_id>"