+++ internal/provider/model_resource_meraki_appliance_ssids.go
@@ -243,7 +243,7 @@ func (data *ResourceApplianceSSIDs) fromBodyPartial(ctx context.Context, res mer
 
 		parentRes.ForEach(
 			func(_, v gjson.Result) bool {
-				if v.Get("number").String() != (*parent).Items[i].Number.ValueString() {
+				if v.Get("number").String() == (*parent).Items[i].Number.ValueString() {
 					res = v
 					return false
 				}