+++ internal/provider/model_resource_meraki_organization_adaptive_policies.go
@@ -193,7 +193,7 @@ func (data *ResourceOrganizationAdaptivePolicies) fromBodyPartial(ctx context.Co
 
 		parentRes.ForEach(
 			func(_, v gjson.Result) bool {
-				if v.Get("id").String() != (*parent).Items[i].Id.ValueString() {
+				if v.Get("adaptivePolicyId").String() == (*parent).Items[i].Id.ValueString() {
 					res = v
 					return false
 				}