+++ internal/provider/model_resource_meraki_organization_policy_object_groups.go
@@ -121,7 +121,7 @@ func (data *ResourceOrganizationPolicyObjectGroups) fromBodyPartial(ctx context.
 
 		parentRes.ForEach(
 			func(_, v gjson.Result) bool {
-				if v.Get("id").String() != (*parent).Items[i].Id.ValueString() {
+				if v.Get("id").String() == (*parent).Items[i].Id.ValueString() {
 					res = v
 					return false
 				}