+++ internal/provider/model_resource_meraki_organization_saml_idps.go
@@ -109,7 +109,7 @@ func (data *ResourceOrganizationSAMLIdPs) fromBodyPartial(ctx context.Context, r
 
 		parentRes.ForEach(
 			func(_, v gjson.Result) bool {
-				if v.Get("id").String() != (*parent).Items[i].Id.ValueString() {
+				if v.Get("idpId").String() == (*parent).Items[i].Id.ValueString() {
 					res = v
 					return false
 				}