+++ internal/provider/model_resource_meraki_sm_admin_roles.go
@@ -121,7 +121,7 @@ func (data *ResourceSMAdminRoles) fromBodyPartial(ctx context.Context, res merak
 
 		parentRes.ForEach(
 			func(_, v gjson.Result) bool {
-				if v.Get("id").String() != (*parent).Items[i].Id.ValueString() {
+				if v.Get("roleId").String() == (*parent).Items[i].Id.ValueString() {
 					res = v
 					return false
 				}