+++ internal/provider/model_resource_meraki_switch_routing_interfaces.go
@@ -209,7 +209,7 @@ func (data *ResourceSwitchRoutingInterfaces) fromBodyPartial(ctx context.Context
 
 		parentRes.ForEach(
 			func(_, v gjson.Result) bool {
-				if v.Get("id").String() != (*parent).Items[i].Id.ValueString() {
+				if v.Get("interfaceId").String() == (*parent).Items[i].Id.ValueString() {
 					res = v
 					return false
 				}